+++ packages/svelte/index.d.ts
@@ -3,6 +3,7 @@ import type { FilterPattern, Plugin } from 'vite'
 export interface Options {
   include?: FilterPattern
   exclude?: FilterPattern
+  preprocess?: boolean
 }
 
 export declare function whyframeSvelte(options?: Options): Plugin